refactor(gui): extract Settings page + save handler into SettingsPageMixin
Move the Settings tab — page builder, _settings_card factory, live
budget-warning updater, and the save handler — off WinpodxWindow
into a private mixin so main_window.py can focus on overall window
orchestration:

- SettingsPageMixin (src/winpodx/gui/_main_window_settings.py)
  _build_settings_page (RDP + container cards + reverse-open panel
  + budget warning + save button), _settings_card (shared card
  factory), _update_budget_warning (live max_sessions × ram_gb
  validator), _save_settings (validate + persist + offer container
  restart if cpu/ram/port/user changed).

WinpodxWindow now reads
`class WinpodxWindow(AppCrudMixin, HeaderMixin, InfoPageMixin,
LibraryPageMixin, LogsMixin, MaintenanceMixin, PodStatusMixin,
SettingsPageMixin, QMainWindow)`.

The reverse-open panel build, the FreeRDP allowlist tooltip, the
container-restart confirmation flow, and the __post_init__ clamp on
max_sessions all come along as-is. No behaviour change; tests:
1096 passed.

main_window.py: 957 -> 627 lines (-330). Cumulative: 2745 -> 627 (-77%).
